Move TradeCalculation.py to its own folder
TradeCalculation.py is just too unwieldy to leave as a single file
any more.  Before I break it up, I'm first moving it to its own
namespace (similarly to DeltaDebug) to avoid cluttering the package
root with too many classes.
