+++ PyRoute/Galaxy.py
@@ -14,7 +14,7 @@ import math
 import networkx as nx
 
 from PyRoute.Star import Star
-from PyRoute.TradeCalculation import TradeCalculation, NoneCalculation, CommCalculation, XRouteCalculation, \
+from PyRoute.Calculation.TradeCalculation import TradeCalculation, NoneCalculation, CommCalculation, XRouteCalculation, \
     OwnedWorldCalculation
 from PyRoute.StatCalculation import ObjectStatistics
 from PyRoute.AllyGen import AllyGen
